# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TradeShield B2B escrow marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in with phone number and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Session started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request an OTP to reset password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordInput"
      responses:
        "200":
          description: OTP sent if account exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpRequestResult"
        "429":
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password using OTP verification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password updated and session started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid or expired code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/request-otp:
    post:
      operationId: requestOtp
      tags: [auth]
      summary: Request an OTP code for registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpRequestInput"
      responses:
        "200":
          description: OTP generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpRequestResult"
        "400":
          description: Invalid phone or account already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limited — wait before resending
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/verify-otp:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify registration OTP before completing signup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpVerifyInput"
      responses:
        "200":
          description: OTP verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpVerifyResult"
        "400":
          description: Invalid or expired code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Complete registration for a phone verified via /auth/verify-otp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Account created and session started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Phone not verified or already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get the currently logged-in user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCurrentUser
      tags: [auth]
      summary: Update the current user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    UserRole:
      type: string
      enum: [buyer, supplier, both]
    User:
      type: object
      required:
        [id, phone, role, businessName, location, isAdmin, kycStatus, createdAt]
      properties:
        id:
          type: integer
        phone:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        businessName:
          type: string
        location:
          type: string
        category:
          type: ["string", "null"]
        payoutMomoNumber:
          type: ["string", "null"]
        isAdmin:
          type: boolean
        kycStatus:
          $ref: "#/components/schemas/KycStatus"
        ghanaCardNumber:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    UserUpdate:
      type: object
      properties:
        businessName:
          type: string
          minLength: 1
        location:
          type: string
          minLength: 1
        category:
          type: string
        payoutMomoNumber:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
...
    OtpRequestInput:
      type: object
      required: [phone]
      properties:
        phone:
          type: string
          minLength: 9
          description: Ghana mobile number (+233 or local 0… format)
    OtpRequestResult:
      type: object
      required: [message]
      properties:
        message:
          type: string
        debugCode:
          type: string
        retryAfterSeconds:
          type: integer
          description: Seconds until another OTP can be requested
    LoginInput:
      type: object
      required: [phone, password]
      properties:
        phone:
          type: string
          minLength: 9
        password:
          type: string
          minLength: 8
    ForgotPasswordInput:
      type: object
      required: [phone]
      properties:
        phone:
          type: string
          minLength: 9
    ResetPasswordInput:
      type: object
      required: [phone, code, password]
      properties:
        phone:
          type: string
        code:
          type: string
          minLength: 6
          maxLength: 6
        password:
          type: string
          minLength: 8
    OtpVerifyInput:
      type: object
      required: [phone, code]
      properties:
        phone:
          type: string
        code:
          type: string
    OtpVerifyResult:
      type: object
      required: [verified, needsRegistration]
      properties:
        verified:
          type: boolean
        needsRegistration:
          type: boolean
        user:
          oneOf:
            - $ref: "#/components/schemas/User"
            - type: "null"
    RegisterInput:
      type: object
      required: [phone, password, businessName, location, role]
      properties:
        phone:
          type: string
        password:
          type: string
          minLength: 8
        businessName:
          type: string
          minLength: 1
        location:
          type: string
          minLength: 1
        role:
          $ref: "#/components/schemas/UserRole"
        category:
          type: string
        payoutMomoNumber:
          type: string
...
    KycStatus:
      type: string
      enum: [none, pending, approved, rejected]
...